Master Upload
This application has been created for windows phone 8.1. The app uses a
SOAP API constructed by our university and a REST API maintained by
LyricWikia. The SOAP API scrapes a list of songs from the BBC Top 40
website. The REST API returns a JSON object containing a URL to the
requested song's lyrics.
The user clicks a song from the list of top 40 tracks and are taken to a
webpage displaying that song's lyrics.